+++ tests/propositional/test_axiom_system.py
@@ -25,7 +25,7 @@ class TestClassicalAxiomSystem(unittest.TestCase):
         self.assertFalse(classical_logic_axiom_system.is_correct_derivation(deriv4, i4))
 
         correct, error_list = classical_logic_axiom_system.is_correct_derivation(deriv4, i4, return_error_list=True)
-        self.assertEqual(error_list[0], CorrectionError(code=ErrorCode.AX_INCORRECT_AXIOM, category="AX", index=0,
+        self.assertEqual(error_list[0], CorrectionError(code=ErrorCode.AX_INCORRECT_AXIOM, index=0,
                                                         description="['→', ['p'], ['→', ['p'], ['q']]] was marked as "
                                                                     "ax1, but is not an instance of that axiom"))
 
@@ -54,7 +54,7 @@ class TestClassicalAxiomSystem(unittest.TestCase):
         self.assertFalse(classical_logic_axiom_system.is_correct_derivation(deriv3, i))
 
         correct, error_list = classical_logic_axiom_system.is_correct_derivation(deriv3, i, return_error_list=True)
-        self.assertEqual(error_list[0], CorrectionError(code=ErrorCode.AX_RULE_INCORRECTLY_APPLIED, category="AX",
+        self.assertEqual(error_list[0], CorrectionError(code=ErrorCode.AX_RULE_INCORRECTLY_APPLIED,
                                                         index=2, description="['→', ['p'], ['p']] was marked as mp, "
                                                                              "but it is not a correct application of "
                                                                              "that rule"))