+++ CHANGELOG.md
@@ -5,6 +5,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.6.2] - 2023-03-17
+### Changed
+- Removed the ``category`` attribute from ``CorrectionError`` (is now inferrable
+  from the numeric error code)
+
 ## [1.6.1] - 2023-03-16
 ### Added
 - Shortcut for importing the classical predicate parser